refactor test_utils structure
