+++ crates/blockifier/src/test_utils.rs
@@ -7,18 +7,15 @@ pub mod initial_test_state;
 pub mod invoke;
 pub mod prices;
 pub mod struct_impls;
+pub mod testing_context;
 
 use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
-use std::sync::Arc;
 
 use cairo_felt::Felt252;
-use cairo_native::starknet::SyscallResult;
 use num_traits::{One, Zero};
-use starknet_api::core::{
-    calculate_contract_address, ClassHash, ContractAddress, EntryPointSelector, Nonce, PatriciaKey,
-};
+use starknet_api::core::{ContractAddress, EntryPointSelector, Nonce, PatriciaKey};
 use starknet_api::deprecated_contract_class::{
     ContractClass as DeprecatedContractClass, EntryPointType,
 };
@@ -27,29 +24,13 @@ use starknet_api::state::StorageKey;
 use starknet_api::transaction::{
     Calldata, ContractAddressSalt, Resource, ResourceBounds, ResourceBoundsMapping,
 };
-use starknet_api::{calldata, class_hash, contract_address, patricia_key, stark_felt};
-use starknet_types_core::felt::Felt;
+use starknet_api::{calldata, contract_address, patricia_key, stark_felt};
 
 use crate::abi::abi_utils::{get_fee_token_var_address, selector_from_name};
 use crate::abi::constants::{self};
-use crate::block_context::BlockContext;
-use crate::execution::call_info::{CallInfo, OrderedEvent};
-use crate::execution::common_hints::ExecutionMode;
-use crate::execution::contract_class::{ContractClass, ContractClassV0, SierraContractClassV1};
-use crate::execution::entry_point::{
-    CallEntryPoint, CallType, ConstructorContext, EntryPointExecutionContext,
-};
-use crate::execution::execution_utils::{execute_deployment, felt_to_stark_felt};
-use crate::execution::sierra_utils::{
-    contract_address_to_felt, felt_to_starkfelt, starkfelt_to_felt,
-};
-use crate::execution::syscalls::hint_processor::{
-    FAILED_TO_CALCULATE_CONTRACT_ADDRESS, FAILED_TO_EXECUTE_CALL,
-};
-use crate::state::cached_state::CachedState;
-use crate::state::state_api::State;
-use crate::test_utils::dict_state_reader::DictStateReader;
-use crate::transaction::objects::AccountTransactionContext;
+use crate::execution::contract_class::{ContractClass, ContractClassV0};
+use crate::execution::entry_point::{CallEntryPoint, CallType};
+use crate::execution::execution_utils::felt_to_stark_felt;
 use crate::utils::const_max;
 
 // TODO(Dori, 1/2/2024): Remove these constants once all tests use the `contracts` and
@@ -409,266 +390,4 @@ pub fn create_calldata(
     calldata.extend(entry_point_args);
 
     Calldata(calldata.into())
-}
-
-pub fn create_custom_deploy_test_state(
-    address_to_class_hash: Vec<(ContractAddress, ClassHash)>,
-    class_hash_to_class: Vec<(ClassHash, ContractClass)>,
-) -> CachedState<DictStateReader> {
-    CachedState::from(DictStateReader {
-        address_to_class_hash: HashMap::from_iter(address_to_class_hash),
-        class_hash_to_class: HashMap::from_iter(class_hash_to_class),
-        ..Default::default()
-    })
-}
-
-pub fn deploy_contract(
-    state: &mut dyn State,
-    class_hash: Felt,
-    contract_address_salt: Felt,
-    calldata: &[Felt],
-) -> SyscallResult<(Felt, Vec<Felt>)> {
-    let deployer_address = ContractAddress::default();
-
-    let class_hash = ClassHash(felt_to_starkfelt(class_hash));
-
-    let wrapper_calldata =
-        Calldata(Arc::new(calldata.iter().map(|felt| felt_to_starkfelt(*felt)).collect()));
-
-    let calculated_contract_address = calculate_contract_address(
-        ContractAddressSalt(felt_to_starkfelt(contract_address_salt)),
-        class_hash,
-        &wrapper_calldata,
-        deployer_address,
-    )
-    .map_err(|_| vec![Felt::from_hex(FAILED_TO_CALCULATE_CONTRACT_ADDRESS).unwrap()])?;
-
-    let ctor_context = ConstructorContext {
-        class_hash,
-        code_address: Some(calculated_contract_address),
-        storage_address: calculated_contract_address,
-        caller_address: deployer_address,
-    };
-
-    let call_info = execute_deployment(
-        state,
-        &mut Default::default(),
-        &mut EntryPointExecutionContext::new(
-            &BlockContext::create_for_testing(),
-            &AccountTransactionContext::Current(Default::default()),
-            ExecutionMode::Execute,
-            false,
-        )
-        .unwrap(),
-        ctor_context,
-        wrapper_calldata,
-        u64::MAX,
-    )
-    .map_err(|_| vec![Felt::from_hex(FAILED_TO_EXECUTE_CALL).unwrap()])?;
-
-    let return_data = call_info.execution.retdata.0.into_iter().map(starkfelt_to_felt).collect();
-    let contract_address_felt = starkfelt_to_felt(*calculated_contract_address.0.key());
-    Ok((contract_address_felt, return_data))
-}
-
-pub fn prepare_erc20_deploy_test_state() -> (ContractAddress, CachedState<DictStateReader>) {
-    ERC20Factory::new().create_state()
-}
-
-pub trait StateFactory {
-    fn get_state(&self) -> CachedState<DictStateReader>;
-
-    fn create_state(&self) -> (ContractAddress, CachedState<DictStateReader>) {
-        let mut state = self.get_state();
-
-        let class_hash = Felt::from_hex(TEST_ERC20_FULL_CONTRACT_CLASS_HASH).unwrap();
-
-        let (contract_address, _) =
-            deploy_contract(&mut state, class_hash, Felt::from(0), self.args().as_slice()).unwrap();
-
-        let contract_address =
-            ContractAddress(PatriciaKey::try_from(felt_to_starkfelt(contract_address)).unwrap());
-
-        (contract_address, state)
-    }
-
-    fn args(&self) -> Vec<Felt> {
-        vec![]
-    }
-}
-
-pub struct ERC20Factory {
-    args: Vec<Felt>,
-    #[allow(dead_code)]
-    state: CachedState<DictStateReader>,
-}
-
-impl ERC20Factory {
-    pub fn new() -> Self {
-        ERC20Factory {
-            args: vec![
-                contract_address_to_felt(Signers::Alice.into()), // Recipient
-                contract_address_to_felt(Signers::Alice.into()), // Owner
-            ],
-            state: create_custom_deploy_test_state(vec![], vec![]),
-        }
-    }
-}
-
-impl StateFactory for ERC20Factory {
-    fn get_state(&self) -> CachedState<DictStateReader> {
-        create_custom_deploy_test_state(
-            vec![(
-                contract_address!(TEST_ERC20_FULL_CONTRACT_ADDRESS),
-                class_hash!(TEST_ERC20_FULL_CONTRACT_CLASS_HASH),
-            )],
-            vec![(
-                class_hash!(TEST_ERC20_FULL_CONTRACT_CLASS_HASH),
-                SierraContractClassV1::from_file(ERC20_FULL_CONTRACT_PATH).into(),
-            )],
-        )
-    }
-
-    fn args(&self) -> Vec<Felt> {
-        self.args.clone()
-    }
-}
-
-pub struct YASERC20Factory {
-    args: Vec<Felt>,
-    #[allow(dead_code)]
-    state: CachedState<DictStateReader>,
-}
-
-impl YASERC20Factory {
-    pub fn new(args: Vec<Felt>) -> Self {
-        YASERC20Factory { args, state: create_custom_deploy_test_state(vec![], vec![]) }
-    }
-}
-
-impl StateFactory for YASERC20Factory {
-    fn get_state(&self) -> CachedState<DictStateReader> {
-        create_custom_deploy_test_state(
-            vec![],
-            vec![(
-                class_hash!(TEST_ERC20_FULL_CONTRACT_CLASS_HASH),
-                SierraContractClassV1::from_file(ERC20_FULL_CONTRACT_PATH).into(),
-            )],
-        )
-    }
-
-    fn args(&self) -> Vec<Felt> {
-        self.args.clone()
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum Signers {
-    Alice,
-    Bob,
-    Charlie,
-}
-
-impl Signers {
-    pub fn get_address(&self) -> ContractAddress {
-        match self {
-            Signers::Alice => ContractAddress(patricia_key!(0x001u128)),
-            Signers::Bob => ContractAddress(patricia_key!(0x002u128)),
-            Signers::Charlie => ContractAddress(patricia_key!(0x003u128)),
-        }
-    }
-}
-
-impl Into<ContractAddress> for Signers {
-    fn into(self) -> ContractAddress {
-        self.get_address()
-    }
-}
-
-impl Into<Felt> for Signers {
-    fn into(self) -> Felt {
-        contract_address_to_felt(self.get_address())
-    }
-}
-
-impl Into<StarkFelt> for Signers {
-    fn into(self) -> StarkFelt {
-        felt_to_starkfelt(contract_address_to_felt(self.get_address()))
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct TestEvent {
-    pub data: Vec<Felt>,
-    pub keys: Vec<Felt>,
-}
-
-impl From<OrderedEvent> for TestEvent {
-    fn from(value: OrderedEvent) -> Self {
-        let event_data = value.event.data.0.iter().map(|e| starkfelt_to_felt(*e)).collect();
-        let event_keys = value.event.keys.iter().map(|e| starkfelt_to_felt(e.0)).collect();
-        Self { data: event_data, keys: event_keys }
-    }
-}
-
-pub struct TestContext {
-    pub contract_address: ContractAddress,
-    pub state: CachedState<DictStateReader>,
-    pub caller_address: ContractAddress,
-    pub events: Vec<TestEvent>,
-}
-
-impl TestContext {
-    pub fn new<T: StateFactory>(state_factory: T) -> Self {
-        let (contract_address, state) = state_factory.create_state();
-        Self { contract_address, state, caller_address: contract_address, events: vec![] }
-    }
-
-    pub fn with_caller(mut self, caller_address: ContractAddress) -> Self {
-        self.caller_address = caller_address;
-        self
-    }
-
-    pub fn call_entry_point(
-        &mut self,
-        entry_point_name: &str,
-        calldata: Vec<StarkFelt>,
-    ) -> Vec<Felt> {
-        let result = self.call_entry_point_raw(entry_point_name, calldata);
-        result.execution.retdata.0.iter().map(|felt| starkfelt_to_felt(*felt)).collect()
-    }
-
-    pub fn call_entry_point_raw(
-        &mut self,
-        entry_point_name: &str,
-        calldata: Vec<StarkFelt>,
-    ) -> CallInfo {
-        let entry_point_selector = selector_from_name(entry_point_name);
-        let calldata = Calldata(Arc::new(calldata));
-
-        let entry_point_call = CallEntryPoint {
-            calldata,
-            entry_point_selector,
-            code_address: Some(self.contract_address),
-            storage_address: self.contract_address,
-            caller_address: self.caller_address,
-            ..external_entry_point(Some(self.contract_address))
-        };
-
-        let result = entry_point_call.execute_directly(&mut self.state).unwrap();
-
-        let events = result.execution.events.clone();
-
-        self.events.extend(events.iter().map(|e| e.clone().into()));
-
-        result
-    }
-
-    pub fn get_event(&self, index: usize) -> Option<TestEvent> {
-        self.events.get(index).cloned()
-    }
-
-    pub fn get_caller(&self) -> ContractAddress {
-        self.caller_address
-    }
 }
